Initial debsync.sh commit.
First working version that can

* Acquire latest rkt release information from github
* Download debian package and signature file
* Verify the signature against the coreos signing key
* Make the package available in a local repository

All downloading and JSON parsing happens in a firejail sandbox for extra
safety. One known issue is that apt complains that it's unable to authenticate
the package upon installation. This is benign in this case as we do perform
validation, but would still be nice to improve.