Provide signed Release file.
Generate a Release file and sign it with a root-owned key. This way, the
repository passes APT authentication so installing the package doesn't require
allowing unauthenticated sources. The key being root-accessible is not a
security issue, attackers with root privileges can add additional trusted APT
keys anyways. It's important to not leak the key though, as that would enable
attackers to produce valid signatures for repositories that are synced via the
net. It'd be nice to bind the key to the local repository so it can't be used
with other repositories, but I don't think that's possible.